backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-dotenv
